changes regarding review
git-svn-id: https://horse.aplana.com/svn/SBRFACCTAX/trunk/taxaccounting@393 34aa7802-8fb0-764d-beb8-9b5b9d595954

+++ core/src/main/java/com/aplana/sbrf/taxaccounting/service/FormDataSearchService.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 public interface FormDataSearchService {
 
-	List<FormData> findDataByUserIdAndFilter(Long userId, FormDataFilter formDataFilter);
+	List<FormData> findDataByUserIdAndFilter(int userId, FormDataFilter formDataFilter);
 
 	List<Department> listDepartments();
